python/open-payments/requirements.txt
requests-http-signature==0.7.1